requirements.txt
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.2.0
django-js-asset==1.2.2
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
asgiref==3.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2.8
django-ckeditor==6.2.0
django-crispy-forms==2.0
django-js-asset==1.2.2
django-modeltranslation==0.18.11
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.3
oauthlib==3.2.2
Pillow==8.4.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2021.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.2
translate==3.6.1
typing_extensions==4.7.1
urllib3==2.0.4
